C3 will not build at this time, it fails to link.  But with so much work done on it, I thought it would be best to add it to source control.

+++ c3/source/win32/ntdll/ntdll_error.c
@@ -56,8 +56,8 @@ ULONG WINAPI RtlNtStatusToDosErrorNoTeb_k( NTSTATUS status )
 
     while (table->start)
     {
-        if (status < table->start) break;
-        if (status < table->end)
+        if ((DWORD)status < table->start) break;
+        if ((DWORD)status < table->end)
         {
             DWORD ret = table->table[status - table->start];
             /* unknown entries are 0 */
